# Conflicts:
#	src/browserist/browser/check_if/__main__.py
#	src/browserist/browser/check_if/contains_text.py
#	src/browserist/browser/check_if/is_image_loaded.py
#	src/browserist/browser/click/__main__.py
#	src/browserist/browser/click/button.py
#	src/browserist/browser/click/button_if_contains_text.py
#	src/browserist/browser/get/__main__.py
#	src/browserist/browser/get/attribute/__main__.py
#	src/browserist/browser/get/attribute/value.py
#	src/browserist/browser/get/attribute/values.py
#	src/browserist/browser/get/dimensions.py
#	src/browserist/browser/get/element.py
#	src/browserist/browser/get/elements.py
#	src/browserist/browser/get/elements_by_tag.py
#	src/browserist/browser/get/text.py
#	src/browserist/browser/get/texts.py
#	src/browserist/browser/get/url/__main__.py
#	src/browserist/browser/get/url/from_image.py
#	src/browserist/browser/get/url/from_images.py
#	src/browserist/browser/get/url/from_link.py
#	src/browserist/browser/get/url/from_links.py
#	src/browserist/browser/input/__main__.py
#	src/browserist/browser/input/clear.py
#	src/browserist/browser/input/value.py
#	src/browserist/browser/mouse/__main__.py
#	src/browserist/browser/mouse/hover.py
#	src/browserist/browser/scroll/__main__.py
#	src/browserist/browser/scroll/into_view.py
#	src/browserist/browser/scroll/into_view_if_not_visible.py
#	src/browserist/browser/select/input_field.py
#	src/browserist/browser/tool/__main__.py
#	src/browserist/browser/tool/count_elements.py
#	src/browserist/browser/wait/__main__.py
#	src/browserist/browser/wait/for_element.py
#	src/browserist/browser/wait/random_time.py
#	src/browserist/browser/wait/until/__main__.py
#	src/browserist/browser/wait/until/element_disappears.py
#	src/browserist/browser/wait/until/images_have_loaded.py
#	src/browserist/browser/wait/until/number_of_window_handles_is.py
#	src/browserist/browser/wait/until/page_title/__main__.py
#	src/browserist/browser/wait/until/page_title/changes.py
#	src/browserist/browser/wait/until/page_title/contains.py
#	src/browserist/browser/wait/until/page_title/equals.py
#	src/browserist/browser/wait/until/text/__main__.py
#	src/browserist/browser/wait/until/text/changes.py
#	src/browserist/browser/wait/until/text/contains.py
#	src/browserist/browser/wait/until/text/equals.py
#	src/browserist/browser/wait/until/url/__main__.py
#	src/browserist/browser/wait/until/url/changes.py
#	src/browserist/browser/wait/until/url/contains.py
#	src/browserist/browser/wait/until/url/equals.py
#	src/browserist/helper/retry.py
#	src/browserist/model/timeout_strategy.py

+++ src/browserist/browser/check_if/__main__.py
@@ -13,7 +13,7 @@ class CheckIfDriverMethods(DriverMethods):
     def __init__(self, browser_driver: BrowserDriver) -> None:
         super().__init__(browser_driver)
 
-    def contains_text(self, xpath: str, regex: str, ignore_case: bool = True, timeout: int | None = None) -> bool:  # type: ignore
+    def contains_text(self, xpath: str, regex: str, ignore_case: bool = True, timeout: float | None = None) -> bool:  # type: ignore
         """Check if element contains text. The condition works for both ordinary text (e.g. "Submit") or regular expression (e.g. r"colou?r"). Note it's a search for text, not a strict text match."""
 
         if self._timeout_should_continue():
@@ -51,7 +51,7 @@ class CheckIfDriverMethods(DriverMethods):
         if self._timeout_should_continue():
             return check_if_is_enabled(self._browser_driver, xpath)
 
-    def is_image_loaded(self, xpath: str, timeout: int | None = None) -> bool:  # type: ignore
+    def is_image_loaded(self, xpath: str, timeout: float | None = None) -> bool:  # type: ignore
         """Check is image is loaded and ready in the DOM."""
 
         if self._timeout_should_continue():
