+++ src/browserist/browser/click/button_if_contains_text.py
@@ -6,7 +6,7 @@ from ..wait.for_element import wait_for_element
 from .button import click_button
 
 
-def click_button_if_contains_text(browser_driver: BrowserDriver, xpath: str, regex: str, ignore_case: bool, timeout: int) -> None:
+def click_button_if_contains_text(browser_driver: BrowserDriver, xpath: str, regex: str, ignore_case: bool, timeout: float) -> None:
     xpath = XPath(xpath)
     wait_for_element(browser_driver, xpath, timeout)
     if check_if_contains_text(browser_driver, xpath, regex, ignore_case, timeout):
