+++ src/browserist/browser/get/__main__.py
@@ -19,7 +19,7 @@ class GetDriverMethods(DriverMethods):
         self.attribute: GetAttributeDriverMethods = GetAttributeDriverMethods(browser_driver)
         self.url: GetUrlDriverMethods = GetUrlDriverMethods(browser_driver)
 
-    def dimensions(self, xpath: str, timeout: int | None = None) -> tuple[int, int]:  # type: ignore
+    def dimensions(self, xpath: str, timeout: float | None = None) -> tuple[int, int]:  # type: ignore
         """Get width and height of element in pixels. Usage:
 
         width, height = browser.get.dimensions("/element/xpath")"""
@@ -28,21 +28,21 @@ class GetDriverMethods(DriverMethods):
             timeout = self._mediate_timeout(timeout)
             return get_dimensions(self._browser_driver, xpath, timeout)
 
-    def element(self, xpath: str, timeout: int | None = None) -> object:  # type: ignore
+    def element(self, xpath: str, timeout: float | None = None) -> object:  # type: ignore
         """Get single web element by XPath."""
 
         if self._timeout_should_continue():
             timeout = self._mediate_timeout(timeout)
             return get_element(self._browser_driver, xpath, timeout)
 
-    def elements(self, xpath: str, timeout: int | None = None) -> list[object]:  # type: ignore
+    def elements(self, xpath: str, timeout: float | None = None) -> list[object]:  # type: ignore
         """Get multiple web elements by XPath."""
 
         if self._timeout_should_continue():
             timeout = self._mediate_timeout(timeout)
             return get_elements(self._browser_driver, xpath, timeout)
 
-    def elements_by_tag(self, tag: str, timeout: int | None = None) -> list[object]:  # type: ignore
+    def elements_by_tag(self, tag: str, timeout: float | None = None) -> list[object]:  # type: ignore
         """"Get multiple web elements by HTML tag. Examples: "img" as tag for all <img> images, "a" for all <a> links, etc."""
 
         if self._timeout_should_continue():
@@ -55,7 +55,7 @@ class GetDriverMethods(DriverMethods):
         if self._timeout_should_continue():
             return get_page_title(self._browser_driver)
 
-    def text(self, xpath: str, timeout: int | None = None) -> str:  # type: ignore
+    def text(self, xpath: str, timeout: float | None = None) -> str:  # type: ignore
         """Get text from element.
 
         This method assumes that the text field shouldn't be empty and therefore will retry to get the text (for better support of single-page apps with extended loading time)."""
@@ -64,7 +64,7 @@ class GetDriverMethods(DriverMethods):
             timeout = self._mediate_timeout(timeout)
             return get_text(self._browser_driver, xpath, timeout)
 
-    def texts(self, xpath: str, timeout: int | None = None) -> list[str]:  # type: ignore
+    def texts(self, xpath: str, timeout: float | None = None) -> list[str]:  # type: ignore
         """Get array of texts from elements.
 
         Assumes that the XPath targets multiple elements."""
