+++ src/browserist/browser/get/attribute/__main__.py
@@ -8,7 +8,7 @@ class GetAttributeDriverMethods(DriverMethods):
     def __init__(self, browser_driver: BrowserDriver) -> None:
         super().__init__(browser_driver)
 
-    def value(self, xpath: str, attribute: str, timeout: int | None = None) -> str:  # type: ignore
+    def value(self, xpath: str, attribute: str, timeout: float | None = None) -> str:  # type: ignore
         """Get value from an attribute of an element. Examples:
 
         Use "src" as attribute to get the source URL from an <img> image tag.
@@ -19,7 +19,7 @@ class GetAttributeDriverMethods(DriverMethods):
             timeout = self._mediate_timeout(timeout)
             return get_attribute_value(self._browser_driver, xpath, attribute, timeout)
 
-    def values(self, xpath: str, attribute: str, timeout: int | None = None) -> list[str]:  # type: ignore
+    def values(self, xpath: str, attribute: str, timeout: float | None = None) -> list[str]:  # type: ignore
         """Get values from an attribute of multiple elements. Assumes that the XPath targets multiple links. Examples:
 
         Use "src" as attribute to get the source URL from an <img> image tag.
