+++ src/browserist/browser/get/attribute/values.py
@@ -5,7 +5,7 @@ from ....model.type.xpath import XPath
 from ...wait.for_element import wait_for_element
 
 
-def get_attribute_values(browser_driver: BrowserDriver, xpath: str, attribute: str, timeout: int) -> list[str]:
+def get_attribute_values(browser_driver: BrowserDriver, xpath: str, attribute: str, timeout: float) -> list[str]:
     xpath = XPath(xpath)
     wait_for_element(browser_driver, xpath, timeout)
     driver = browser_driver.get_webdriver()
