+++ src/browserist/browser/get/dimensions.py
@@ -5,7 +5,7 @@ from ...model.type.xpath import XPath
 from ..wait.for_element import wait_for_element
 
 
-def get_dimensions(browser_driver: BrowserDriver, xpath: str, timeout: int) -> tuple[int, int]:
+def get_dimensions(browser_driver: BrowserDriver, xpath: str, timeout: float) -> tuple[int, int]:
     xpath = XPath(xpath)
     wait_for_element(browser_driver, xpath, timeout)
     driver = browser_driver.get_webdriver()
