+++ src/browserist/browser/get/elements_by_tag.py
@@ -4,7 +4,7 @@ from ...model.browser.base.driver import BrowserDriver
 from ..wait.for_element import wait_for_element
 
 
-def get_elements_by_tag(browser_driver: BrowserDriver, tag: str, timeout: int) -> list[object]:
+def get_elements_by_tag(browser_driver: BrowserDriver, tag: str, timeout: float) -> list[object]:
     wait_for_element(browser_driver, f"//{tag}", timeout)
     driver = browser_driver.get_webdriver()
     return driver.find_elements(By.TAG_NAME, tag)  # type: ignore
