+++ src/browserist/browser/get/text.py
@@ -6,7 +6,7 @@ from ...model.type.xpath import XPath
 from ..wait.for_element import wait_for_element
 
 
-def get_text(browser_driver: BrowserDriver, xpath: str, timeout: int) -> str:
+def get_text(browser_driver: BrowserDriver, xpath: str, timeout: float) -> str:
     def get_inner_text_of_element(browser_driver: BrowserDriver, xpath: str) -> str:
         return browser_driver.webdriver.find_element(By.XPATH, xpath).text  # type: ignore
