+++ src/browserist/browser/input/value.py
@@ -5,7 +5,7 @@ from ...model.type.xpath import XPath
 from ..wait.for_element import wait_for_element
 
 
-def input_value(browser_driver: BrowserDriver, xpath: str, value: str, timeout: int) -> None:
+def input_value(browser_driver: BrowserDriver, xpath: str, value: str, timeout: float) -> None:
     xpath = XPath(xpath)
     wait_for_element(browser_driver, xpath, timeout)
     driver = browser_driver.get_webdriver()
