+++ src/browserist/browser/mouse/hover.py
@@ -6,7 +6,7 @@ from ...model.type.xpath import XPath
 from ..wait.for_element import wait_for_element
 
 
-def mouse_hover(browser_driver: BrowserDriver, xpath: str, timeout: int) -> None:
+def mouse_hover(browser_driver: BrowserDriver, xpath: str, timeout: float) -> None:
     xpath = XPath(xpath)
     wait_for_element(browser_driver, xpath, timeout)
     driver = browser_driver.get_webdriver()
