+++ src/browserist/browser/scroll/__main__.py
@@ -32,14 +32,14 @@ class ScrollDriverMethods(DriverMethods):
         if self._timeout_should_continue():
             scroll_down_by(self._browser_driver, pixels, delay_seconds)
 
-    def into_view(self, xpath: str, timeout: int | None = None, delay_seconds: float = 1) -> None:
+    def into_view(self, xpath: str, timeout: float | None = None, delay_seconds: float = 1) -> None:
         """Find element and scroll up or down until element is visible. Add custom delay in seconds to ensure the view is updated after scroll."""
 
         if self._timeout_should_continue():
             timeout = self._mediate_timeout(timeout)
             scroll_into_view(self._browser_driver, xpath, timeout, delay_seconds)
 
-    def into_view_if_not_visible(self, xpath: str, timeout: int | None = None, delay_seconds: float = 1) -> None:
+    def into_view_if_not_visible(self, xpath: str, timeout: float | None = None, delay_seconds: float = 1) -> None:
         """If not visible, find element and scroll up or down until element is visible. Add custom delay in seconds to ensure the view is updated after scroll."""
 
         if self._timeout_should_continue():
