+++ src/browserist/browser/select/input_field.py
@@ -5,7 +5,7 @@ from ...model.type.xpath import XPath
 from ..wait.for_element import wait_for_element
 
 
-def select_input_field(browser_driver: BrowserDriver, xpath: str, timeout: int) -> None:
+def select_input_field(browser_driver: BrowserDriver, xpath: str, timeout: float) -> None:
     xpath = XPath(xpath)
     wait_for_element(browser_driver, xpath, timeout)
     driver = browser_driver.get_webdriver()
