+++ src/browserist/browser/tool/count_elements.py
@@ -5,7 +5,7 @@ from ...model.type.xpath import XPath
 from ..wait.for_element import wait_for_element
 
 
-def tool_count_elements(browser_driver: BrowserDriver, xpath: str, timeout: int) -> int:
+def tool_count_elements(browser_driver: BrowserDriver, xpath: str, timeout: float) -> int:
     xpath = XPath(xpath)
     wait_for_element(browser_driver, xpath, timeout)
     driver = browser_driver.get_webdriver()
