+++ src/browserist/browser/wait/__main__.py
@@ -12,7 +12,7 @@ class WaitDriverMethods(DriverMethods):
         super().__init__(browser_driver)
         self.until: WaitUntilDriverMethods = WaitUntilDriverMethods(browser_driver)
 
-    def for_element(self, xpath: str, timeout: int | None = None) -> None:
+    def for_element(self, xpath: str, timeout: float | None = None) -> None:
         """Wait until element is ready in the DOM and/or on the screen.
 
         Especially useful for single-page app elements handled/modified by JavaScript, but also standard HTML that doesn't load immediately, this helper function ensures that DOM elements are ready before processing."""
@@ -21,7 +21,7 @@ class WaitDriverMethods(DriverMethods):
             timeout = self._mediate_timeout(timeout)
             wait_for_element(self._browser_driver, xpath, timeout)
 
-    def random_time(self, min_seconds: int = 1, max_seconds: int = 5) -> None:
+    def random_time(self, min_seconds: float = 1, max_seconds: float = 5) -> None:
         """Randomize sleep timing to make actions look less like a bot."""
 
         if self._timeout_should_continue():
