+++ src/browserist/browser/wait/until/number_of_window_handles_is.py
@@ -7,7 +7,7 @@ from ....exception.timeout import WaitForWindowTimeoutException
 from ....model.browser.base.driver import BrowserDriver
 
 
-def wait_until_number_of_window_handles_is(browser_driver: BrowserDriver, expected_handles: int, timeout: int) -> None:
+def wait_until_number_of_window_handles_is(browser_driver: BrowserDriver, expected_handles: int, timeout: float) -> None:
     try:
         driver = browser_driver.get_webdriver()
         WebDriverWait(driver, timeout).until(EC.number_of_windows_to_be(expected_handles))  # type: ignore
