+++ src/browserist/browser/wait/until/page_title/__main__.py
@@ -9,21 +9,21 @@ class WaitUntilPageTitleDriverMethods(DriverMethods):
     def __init__(self, browser_driver: BrowserDriver) -> None:
         super().__init__(browser_driver)
 
-    def changes(self, baseline_text: str, timeout: int | None = None) -> None:
+    def changes(self, baseline_text: str, timeout: float | None = None) -> None:
         """Wait until the page title changes from a baseline text, e.g. after a page reload or change. The text is evaluated as an exact match."""
 
         if self._timeout_should_continue():
             timeout = self._mediate_timeout(timeout)
             wait_until_page_title_changes(self._browser_driver, baseline_text, timeout)
 
-    def contains(self, page_title_fragment: str, timeout: int | None = None) -> None:
+    def contains(self, page_title_fragment: str, timeout: float | None = None) -> None:
         """Wait until the page title has changed, e.g. after a redirect or update. The input can contain both a fragment or the full page title."""
 
         if self._timeout_should_continue():
             timeout = self._mediate_timeout(timeout)
             wait_until_page_title_contains(self._browser_driver, page_title_fragment, timeout)
 
-    def equals(self, page_title: str, timeout: int | None = None) -> None:
+    def equals(self, page_title: str, timeout: float | None = None) -> None:
         """Wait until the page title has changed, e.g. after a redirect or update. The input has to match the exact page title."""
 
         if self._timeout_should_continue():
