+++ src/browserist/browser/wait/until/page_title/changes.py
@@ -3,7 +3,7 @@ from .....model.browser.base.driver import BrowserDriver
 from ....get.page_title import get_page_title
 
 
-def wait_until_page_title_changes(browser_driver: BrowserDriver, baseline_text: str, timeout: int) -> None:
+def wait_until_page_title_changes(browser_driver: BrowserDriver, baseline_text: str, timeout: float) -> None:
     def has_page_title_changed(browser_driver: BrowserDriver, baseline_text: str) -> bool:
         return get_page_title(browser_driver) != baseline_text
