+++ src/browserist/browser/wait/until/page_title/contains.py
@@ -7,7 +7,7 @@ from .....exception.timeout import WaitForPageTitleToChangeTimeoutException
 from .....model.browser.base.driver import BrowserDriver
 
 
-def wait_until_page_title_contains(browser_driver: BrowserDriver, page_title_fragment: str, timeout: int) -> None:
+def wait_until_page_title_contains(browser_driver: BrowserDriver, page_title_fragment: str, timeout: float) -> None:
     try:
         driver = browser_driver.get_webdriver()
         WebDriverWait(driver, timeout).until(EC.title_contains(page_title_fragment))  # type: ignore
