+++ src/browserist/browser/wait/until/page_title/equals.py
@@ -7,7 +7,7 @@ from .....exception.timeout import WaitForPageTitleToChangeTimeoutException
 from .....model.browser.base.driver import BrowserDriver
 
 
-def wait_until_page_title_equals(browser_driver: BrowserDriver, page_title: str, timeout: int) -> None:
+def wait_until_page_title_equals(browser_driver: BrowserDriver, page_title: str, timeout: float) -> None:
     try:
         driver = browser_driver.get_webdriver()
         WebDriverWait(driver, timeout).until(EC.title_is(page_title))  # type: ignore
