+++ src/browserist/browser/wait/until/text/__main__.py
@@ -9,21 +9,21 @@ class WaitUntilTextDriverMethods(DriverMethods):
     def __init__(self, browser_driver: BrowserDriver) -> None:
         super().__init__(browser_driver)
 
-    def changes(self, xpath: str, baseline_text: str, timeout: int | None = None) -> None:
+    def changes(self, xpath: str, baseline_text: str, timeout: float | None = None) -> None:
         """Wait until the text of an element changes from a baseline text, e.g. after a form action. The text is evaluated as an exact match."""
 
         if self._timeout_should_continue():
             timeout = self._mediate_timeout(timeout)
             wait_until_text_changes(self._browser_driver, xpath, baseline_text, timeout)
 
-    def contains(self, xpath: str, regex: str, timeout: int | None = None) -> None:
+    def contains(self, xpath: str, regex: str, timeout: float | None = None) -> None:
         """Wait until the text of an element has changed, e.g. after a form action."""
 
         if self._timeout_should_continue():
             timeout = self._mediate_timeout(timeout)
             wait_until_text_contains(self._browser_driver, xpath, regex, timeout)
 
-    def equals(self, xpath: str, regex: str, timeout: int | None = None) -> None:
+    def equals(self, xpath: str, regex: str, timeout: float | None = None) -> None:
         """Wait until the text of an element has changed, e.g. after a form action. The text is evaluated as an exact match."""
 
         if self._timeout_should_continue():
