+++ src/browserist/browser/wait/until/url/changes.py
@@ -4,7 +4,7 @@ from .....model.type.url import URL
 from ....get.url.current import get_current_url
 
 
-def wait_until_url_changes(browser_driver: BrowserDriver, baseline_url: str, timeout: int) -> None:
+def wait_until_url_changes(browser_driver: BrowserDriver, baseline_url: str, timeout: float) -> None:
     def has_url_changed(browser_driver: BrowserDriver, baseline_url: str) -> bool:
         return get_current_url(browser_driver) != baseline_url
